Add generic simplex solve header, move existing primal solve to own cpp
